+++ src/main/java/com/trego/service/impl/OrderServiceImpl.java
@@ -741,10 +741,10 @@ public class OrderServiceImpl implements IOrderService {
                 
                 Medicine medicine = null;
                 if (orderItem.getVendorMedicineId() != null) {
-                    medicine = medicineRepository.findById(orderItem.getVendorMedicineId()).orElse(null);
+                    medicine = medicineRepository.findById(orderItem.getVendorMedicineId().intValue()).orElse(null);
                 }
                 if (medicine == null && orderItem.getMedicineId() != null) {
-                    medicine = medicineRepository.findById(orderItem.getMedicineId()).orElse(null);
+                    medicine = medicineRepository.findById(orderItem.getMedicineId().intValue()).orElse(null);
                 }
 
                 String name = "";
@@ -1138,7 +1138,7 @@ public class OrderServiceImpl implements IOrderService {
 
     private MedicineDTO populateMedicalDTO(MedicineDTO medicineDTO, Stock stock) {
 
-        Medicine tempMedicine = medicineRepository.findById(medicineDTO.getId()).orElse(null);
+        Medicine tempMedicine = medicineRepository.findById((int) medicineDTO.getId()).orElse(null);
         if (tempMedicine == null) return medicineDTO;
         
         // Retrieve master medicine to serve as catalog fallback
@@ -1567,7 +1567,7 @@ public class OrderServiceImpl implements IOrderService {
     }
 
     private Medicine resolveMedicine(long medicineId, Integer vendorId) {
-        Medicine med = medicineRepository.findById(Long.valueOf(medicineId)).orElse(null);
+        Medicine med = medicineRepository.findById((int) medicineId).orElse(null);
         if (med == null && vendorId != null) {
             Vendor vendor = vendorRepository.findById(vendorId).orElse(null);
             if (vendor == null) {
